Created Player detection functionality
The enemy AI will now detect the player when walked in front of.  They will follow the player until the player remains undetected for 3 seconds and the enemy will return to their patrol route.  Added a minor edit to the character health script to allow enemy game objects to be destroyed on death so they do not continue to raycast and run coroutines.